refactor: useGetCategories 파일명 오타 수정
