+++ src/api/hooks/categories.mock.ts
@@ -1,6 +1,6 @@
 import { rest } from 'msw';
 
-import { getCategoriesPath } from './useGetCategorys';
+import { getCategoriesPath } from './useGetCategories';
 
 export const categoriesMockHandler = [
   rest.get(getCategoriesPath(), (_, res, ctx) => {
