+++ src/hooks/useCurrentCategory.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 
-import { useGetCategories } from '@/api/hooks/useGetCategorys';
+import { useGetCategories } from '@/api/hooks/useGetCategories';
 import { getCurrentCategory } from '@/components/features/Category/CategoryHeroSection';
 
 type Props = { categoryId: string };
